class name updated to SubChildDoesNotAccessParentMethod
